+++ geom/aabb.hpp
@@ -58,12 +58,12 @@ namespace util { namespace geom {
     };
 
     typedef AABB<2,float> AABB2f;
-    typedef AABB<2,size_t> AABB2u;
-    typedef AABB<2,intmax_t> AABB2i;
+    typedef AABB<2,unsigned> AABB2u;
+    typedef AABB<2,int> AABB2i;
 
     typedef AABB<3,float> AABB3f;
-    typedef AABB<3,size_t> AABB3u;
-    typedef AABB<3,intmax_t> AABB3i;
+    typedef AABB<3,unsigned> AABB3u;
+    typedef AABB<3,int> AABB3i;
 } }
 
 #include "aabb.ipp"
